+++ src/ds/network/packet_chunker.cpp
@@ -13,19 +13,19 @@ Chunker::Chunker()
 {
 }
 
-void Chunker::Chunkify(const char *src, size_t size, size_t groupId, std::vector<std::string> &dst){
+void Chunker::Chunkify(const char *src, unsigned size, unsigned groupId, std::vector<std::string> &dst){
 
-	size_t chunkSize = mChunkSize - sizeof(ChunkHeader);
-	size_t pos = 0;
+	unsigned chunkSize = mChunkSize - sizeof(ChunkHeader);
+	unsigned pos = 0;
 
 	snappy::Compress(src, size, &mCompressedBuffer);
 
-	size_t nSize = mCompressedBuffer.size();
-	size_t numIterations = nSize / chunkSize;
+	unsigned nSize = mCompressedBuffer.size();
+	unsigned numIterations = nSize / chunkSize;
 
-	size_t excess = nSize % chunkSize;
+	unsigned excess = nSize % chunkSize;
 
-	size_t total = numIterations + (excess ? 1 : 0);
+	unsigned total = numIterations + (excess ? 1 : 0);
 
 	dst.resize(total);
 
@@ -49,7 +49,7 @@ void Chunker::Chunkify(const char *src, size_t size, size_t groupId, std::vector
 	}
 }
 
-void Chunker::Chunkify(std::string src, size_t groupId, std::vector<std::string> &dst){
+void Chunker::Chunkify(std::string src, unsigned groupId, std::vector<std::string> &dst){
 	Chunkify(src.c_str(), src.size(), groupId, dst);
 }
 
@@ -62,7 +62,7 @@ bool DeChunker::addChunk(std::string &chunk){
 	return addChunk(chunk.c_str(), chunk.size());
 }
 
-bool DeChunker::addChunk(const char *chunk, size_t size){
+bool DeChunker::addChunk(const char *chunk, unsigned size){
 	if(size < sizeof(ChunkHeader)){
 		return false;
 	}
@@ -106,7 +106,7 @@ bool DeChunker::addChunk(const char *chunk, size_t size){
 			mReceived.pop_front();
 		}
 
-		std::sort(mGroupsReceived.begin(), mGroupsReceived.end(), [](size_t a, size_t b) -> bool {
+		std::sort(mGroupsReceived.begin(), mGroupsReceived.end(), [](unsigned a, unsigned b) -> bool {
 			if(a > b)
 				return true;
 			return false;
@@ -117,7 +117,7 @@ bool DeChunker::addChunk(const char *chunk, size_t size){
 
 	if(stats.mIdsMissing.empty()){
 		mGroupsAvailable.push_back(chunkHeader.mGroupId);
-		std::sort(mGroupsAvailable.begin(), mGroupsAvailable.end(), [](size_t a, size_t b) -> bool {
+		std::sort(mGroupsAvailable.begin(), mGroupsAvailable.end(), [](unsigned a, unsigned b) -> bool {
 			if(a > b){
 				return true;
 			}
@@ -128,7 +128,7 @@ bool DeChunker::addChunk(const char *chunk, size_t size){
 	return false;
 }
 
-void DeChunker::addChunkToGroup(DeChunkStats &stats, const char *chunk, size_t size){
+void DeChunker::addChunkToGroup(DeChunkStats &stats, const char *chunk, unsigned size){
 	ChunkHeader chunkHeader;
 	memcpy(reinterpret_cast<char *>(&chunkHeader), chunk, sizeof(ChunkHeader));
 
@@ -136,8 +136,8 @@ void DeChunker::addChunkToGroup(DeChunkStats &stats, const char *chunk, size_t s
 	if(found == stats.mIdsMissing.end())
 		return;
 
-	size_t nSize = chunkHeader.mSize / chunkHeader.mChunkSize;
-	size_t pos = chunkHeader.mId * chunkHeader.mChunkSize;
+	unsigned nSize = chunkHeader.mSize / chunkHeader.mChunkSize;
+	unsigned pos = chunkHeader.mId * chunkHeader.mChunkSize;
 
 	if(stats.mData.get()){
 		std::copy(chunk + sizeof(ChunkHeader), chunk + size, stats.mData.get()->begin() + pos);
@@ -148,7 +148,7 @@ void DeChunker::addChunkToGroup(DeChunkStats &stats, const char *chunk, size_t s
 bool DeChunker::getNextGroup(std::string &dst){
 	if(!mGroupsAvailable.empty() && !mGroupsReceived.empty())	{
 
-		size_t groupId = mGroupsAvailable.back();
+		unsigned groupId = mGroupsAvailable.back();
 		
 		auto gr = std::find(mGroupsReceived.begin(), mGroupsReceived.end(), groupId);
 		if(gr != mGroupsReceived.end()){
